fix brand new files not getting parsed
this was a weird one. visual studio code, when saving a new file, does *not* open the file with all the content you saved in it. instead, it opens the file with an empty string, which is weird, but ok. Gapbuffer was mad when the length of the string was zero becuase we used &buffer[0] for getting the pointer, but that breaks ABC. So instead, switching to use the data() ptr works fine. Anyway, why does vscode do that?